+++ cgm_07/filereader.h
@@ -31,7 +31,7 @@ class FileReader
     ~FileReader();
     bool hasLine();
     std::string getString();
-    float getFloat();
+    double getDouble();
     Vector2 getVector2();
     Vector3 getVector3();
