+++ cgm_10/filereader.cpp
@@ -50,23 +50,23 @@ std::string FileReader::getString()
     return str;
 }
 
-float FileReader::getFloat()
+double FileReader::getDouble()
 {
     std::string p1;
     lineStream >> p1;
-    return stof(p1);
+    return stod(p1);
 }
 
 Vector2 FileReader::getVector2()
 {
     std::string p1, p2;
     lineStream >> p1 >> p2;
-    return (Vector2){stof(p1), stof(p2)};
+    return (Vector2){stod(p1), stod(p2)};
 }
 
 Vector3 FileReader::getVector3()
 {
     std::string p1, p2, p3;
     lineStream >> p1 >> p2 >> p3;
-    return (Vector3){stof(p1), stof(p2), stof(p3)};
+    return (Vector3){stod(p1), stod(p2), stod(p3)};
 }
